test3

+++ yuku.js
@@ -460,7 +460,7 @@ message.author.send(`
 \`${p2}печаль\`
 \`${p2}курить\``)
 }
-/*if (message.content.startsWith(p + `профиль`)) { 
+if (message.content.startsWith(p + `профиль`)) { 
     message.delete()
  //   let aicon = message.author.iconURL;
   //  let sjoin = message.member.joinedAt;
@@ -473,7 +473,7 @@ message.author.send(`
   Посмотерть свой счёт: \`;bal\`
   Посмотреть свой уровень: \`!rank\``)
   message.channel.send({embed});
-}*/                                         
+}                                         
 ////////////////////////////////////////////////////////////////////////////////
 if (message.content.startsWith(p + `news`) && (message.author.id === "406343162651738112" || message.author.id === "341988428457705482")) {
     let say = message.content.slice((p + 'embed').length);
